requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bottle==0.13.2
click==8.1.8
fastapi==0.115.8
h11==0.14.0
idna==3.10
proxy_tools==0.1.0
pycairo==1.27.0
pydantic==2.10.6
pydantic_core==2.27.2
PyGObject==3.50.0
pywebview==5.4
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
